requirements.txt
python-dotenv==1.0.0
numpy==1.24.3
Flask==3.0.0
werkzeug==3.0.1
tensorflow==2.13.0
Pillow==10.1.0
google-cloud-storage==2.14.0